backend/transacoes_service/requirements.txt
fastapi
uvicorn
pyodbc
requests
pyjwt
